+++ src/pages/chatroom/index.tsx
@@ -3,10 +3,11 @@ import MainLayout from '@/components/layout/MainLayout';
 import Modal from '@/components/ui/Modal';
 import QuestionBar from '@/components/ui/QuestionBar';
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { useState, useEffect, useRef } from 'react';
+import { useState } from 'react';
 import { useRouter } from 'next/router';
 import { get } from '@/libs/api';
 import { useQueryClient } from '@tanstack/react-query';
+import useReversedInfiniteScroll from '@/hooks/common/useReversedInfiniteScroll';
 
 type Letters = {
   letters: LetterType[];
@@ -45,86 +46,17 @@ const getLetters = async ({ pageParam }: PageParam) => {
   const path = '/api/v1/letters';
   const apiEndpoint = `${baseURL}${path}`;
 
-  console.log('pageparam: ', pageParam);
   const url = pageParam
     ? `${apiEndpoint}?next-cursor=${pageParam}`
     : apiEndpoint;
 
   try {
     const response = await get(url);
-    console.log('response.data: ', response.data);
     return response.data;
   } catch (error) {
     throw error;
   }
 };
-
-// useReversedInfiniteScroll라는 커스텀 훅
-// containerRef과 handleScroll을 뽑아내기 위해 만들었다.
-function useReversedInfiniteScroll(
-  fetchNextPage: () => Promise<unknown>,
-  hasNextPage: boolean = false,
-  dataLength: number = 0
-) {
-  const containerRef = useRef<HTMLDivElement>(null);
-  const isLoadingRef = useRef(false);
-  const prevScrollHeight = useRef(0);
-
-  // useEffect(함수, 배열)
-  // 맨 처음에, prevScrollHeight가 바뀔 때, dataLength가 바뀔 때 실행된다.
-  useEffect(
-    // 스크롤 위치를 지정하는 함수(scrollTo로 이동)
-    function scrollToPrevScrollHeight() {
-      // scrollHeight: 스크롤 된 내용의 전체 높이
-      // 처음 접속했을 때 스크롤을 맨 아래로 내린다.
-      if (prevScrollHeight.current === 0) {
-        containerRef.current?.scrollTo({
-          top: containerRef.current.scrollHeight,
-        });
-        return;
-      }
-
-      // **중요**: 아래 코드를 쓰기 위해 useReversedInfiniteScroll 훅이 존재한다.
-      // IntersectionObserver를 처음에 사용했었음. 근데 이 코드가 없어서 페이지 추가시에 스크롤이 맨 위로 올라가는 문제가 발생했다.
-      // 처음이 아니면 화면을 이동한 만큼만 스크롤 위치를 이동한다.
-      // ㄴ 아래 코드가 없으면 페이지가 바뀔 때 스크롤 위치가 유지되지 않음
-      // ㄴ 그래서 페이지가 생겼을 때 스크롤이 바로 최상단으로 이동하는 문제가 발생한다.
-      containerRef.current?.scrollTo({
-        top: containerRef.current.scrollHeight - prevScrollHeight.current,
-      });
-    },
-    [prevScrollHeight, dataLength]
-  );
-
-  // 스크롤 이벤트 핸들러
-  async function handleScroll() {
-    // 만약 다음 페이지가 없다면
-    if (!hasNextPage) {
-      // 코드를 실행하지 않습니다.
-      return;
-    }
-
-    // scrollHeight은 동일
-    // scrollTop은 스크롤이 올라갈수록 줄어듦
-    // 즉 prevScrollHeight는 스크롤을 올릴 수록 커짐
-    prevScrollHeight.current =
-      (containerRef.current?.scrollHeight ?? 0) -
-      (containerRef.current?.scrollTop ?? 0);
-
-    // 만약 스크롤이 450px 미만으로 줄어들었다면(맨 위까지 스크롤이 올라가기 전에) 다음 페이지 가져오기
-    if ((containerRef.current?.scrollTop ?? 0) < 450 && !isLoadingRef.current) {
-      isLoadingRef.current = true;
-      await fetchNextPage();
-      isLoadingRef.current = false;
-    }
-  }
-
-  // 이 훅은 containerRef과 handleScroll을 리턴한다.
-  return {
-    containerRef,
-    handleScroll,
-  };
-}
 
 const Page: NextPageWithLayout<Letters> = () => {
   const queryClient = useQueryClient();
@@ -147,7 +79,6 @@ const Page: NextPageWithLayout<Letters> = () => {
     // ㄴ 리턴값이 undefined라면 더 이상 가져올 페이지가 없다는 뜻
     // 백엔드에서 넘겨주는 nextCursor: 더 이상 가져올 페이지가 없을 때 -1임
     getNextPageParam: (lastPage) => {
-      console.log('lastPage.nextCursor: ', lastPage.nextCursor);
       return lastPage.nextCursor === -1 ? undefined : lastPage.nextCursor;
     },
 
@@ -156,7 +87,6 @@ const Page: NextPageWithLayout<Letters> = () => {
     // data.pages.reverse()를 하면 앞뒤로 새 페이지가 붙는 버그가 생긴다.
     // ㄴ reverse()는 배열 자체를 바꿔버리기 때문. 대신 [...data.pages].reverse()를 해줘야 한다.
     select: (data) => {
-      // console.log('...data.pages: ', ...data.pages);
       return ([...data.pages] ?? []).reverse().flatMap((page) => page.letters);
     },
   });
@@ -168,8 +98,6 @@ const Page: NextPageWithLayout<Letters> = () => {
     hasNextPage,
     data?.length ?? 0 // 훅에서 dataLength가 됨
   );
-  console.log('hasNextPage: ', hasNextPage);
-  console.log(data);
   const handleQuestionBarClick = ({ letter }: { letter: LetterType }) => {
     if (letter.answerCount === 0) {
       setModalInfo({
@@ -241,9 +169,6 @@ const Page: NextPageWithLayout<Letters> = () => {
           );
         })}
       </div>
-      {/* <button onClick={() => fetchNextPage()}>
-        {hasNextPage ? '더 로드하기' : '더 로드할 것이 없음!'}
-      </button> */}
     </div>
   );
 };
